rovodev pre phase 3: complete architectural hardening - all priorities addressed
Priority 1: Eliminated direct database access from blueprints
- Fixed AI blueprint violations by moving logic to DocumentService
- Fixed health blueprint by using health check utilities
- All blueprints now properly delegate to service layer

Priority 2: Moved business logic from app.py to services
- perform_checklist_ai_analysis -> DocumentService
- would_create_circular_dependency -> TaskService
- check_and_update_project_completion -> ProjectService
- Clean application factory pattern achieved

Priority 3: Standardized transaction management
- AdminService converted to instance methods with @transactional
- Removed all manual db.session.commit/rollback calls
- Consistent transaction handling across all services

Priority 4: Housekeeping completed
- Deleted obsolete events/schemas_old.py
- Added production warning to SECRET_KEY configuration

All architectural violations resolved - ready for Phase 3
